fixed serial to decimal so user can dictate session ID to match cart Id

+++ store/src/scripts/createDBTables.sql
@@ -11,7 +11,7 @@ CREATE TABLE products(itemId SERIAL, name VARCHAR(255), msrp DECIMAL(6,2), saleP
 CREATE TABLE carts(id SERIAL, username VARCHAR(255), completed DECIMAL, UNIQUE(id));
 
 
-CREATE TABLE sessions(id SERIAL, productId DECIMAL);
+CREATE TABLE sessions(id DECIMAL, productId DECIMAL);
 
 
 SELECT "You will need to decide how to persist cart and order information. Will you amend an existing table or create new one(s)?" AS "";
